Add Filament Check Whois Widget plugin
The commit introduces a new Filament plugin named "Filament Check Whois Widget". It consists of several components- a command, a facade, and an associated service provider. Additionally, it provides configuration and language files, a database migration template and the plugin register itself. Meanwhile, this commit removes the initial skeleton files that were no longer necessary for the functioning of the application.